Clean player and team controller
The disconnect method now calls player stored in PlayerController global
object for now. The problem is the use of the `this` keyword that Tim
uses. It's refering to the constructor object itself instead o the the
PlayerController object we're exporting.
Also fix the set_initials' promise object chaining.

The team controller also had a mistyped method.